Rename updater user-facing branding and target executable

+++ UndertaleModToolUpdater/Program.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
@@ -10,33 +10,27 @@ namespace UndertaleModToolUpdater
     {
         static void Main(string[] args)
         {
-            Console.Title = "UndertaleModTool updater";
+            Console.Title = "PizzaTowerModTool updater";
 
-            // Get all processes named UndertaleModTool
-            Process[] utmtInstances = Process.GetProcessesByName("UndertaleModTool");
-            
-            if (utmtInstances.Length > 0)
+            Process[] instances = Process.GetProcessesByName("PizzaTowerModTool");
+            if (instances.Length > 0)
             {
-                // Found some! Wait for all of them to close
-                Console.WriteLine("Waiting for UndertaleModTool to close...");
-                foreach (var instance in utmtInstances)
-                {
+                Console.WriteLine("Waiting for PizzaTowerModTool to close...");
+                foreach (var instance in instances)
                     instance.WaitForExit();
-                }
             }
 
-            string basePath = Path.Join(Path.GetTempPath(), "UndertaleModTool") + Path.DirectorySeparatorChar;
+            string basePath = Path.Join(Path.GetTempPath(), "PizzaTowerModTool") + Path.DirectorySeparatorChar;
             string appPath = null;
 
-            // Check if Update.zip actually exists
-            if (!File.Exists(basePath + "Update.zip")) {
-                Console.WriteLine("Update.zip is missing! This program is not meant to be ran by itself, please update through UndertaleModTool.");
+            if (!File.Exists(basePath + "Update.zip"))
+            {
+                Console.WriteLine("Update.zip is missing! This program is not meant to be run by itself; please update through PizzaTowerModTool.");
                 Console.WriteLine("Press any key to exit...");
                 Console.Read();
                 Environment.Exit(1);
             }
 
-            // If this exists from a failed update or something, then remove it
             if (Directory.Exists(basePath + "Update"))
             {
                 Console.WriteLine("Removing Update folder...");
@@ -50,53 +44,40 @@ namespace UndertaleModToolUpdater
                 Console.Read();
                 Environment.Exit(1);
             }
-            else
-            {
-                appPath = File.ReadAllText("actualAppFolder");
-                File.Delete("actualAppFolder");
-            }
 
-            // Extract the update ZIP
+            appPath = File.ReadAllText("actualAppFolder");
+            File.Delete("actualAppFolder");
+
             Console.WriteLine("Extracting Update.zip...");
             ZipFile.ExtractToDirectory(basePath + "Update.zip", basePath + "Update", true);
             Console.WriteLine("Deleting Update.zip...");
             File.Delete(basePath + "Update.zip");
-
             Console.WriteLine("Replacing files with update...");
             MoveDirectory(basePath + "Update", appPath);
+            Console.WriteLine("Finished updating, launching PizzaTowerModTool...");
 
-            Console.WriteLine("Finished updating, launching UTMT...");
-
-            Process.Start(new ProcessStartInfo(Path.Join(appPath, "UndertaleModTool.exe"))
+            Process.Start(new ProcessStartInfo(Path.Join(appPath, "PizzaTowerModTool.exe"))
             {
                 WorkingDirectory = appPath,
                 Arguments = "deleteTempFolder"
             });
-            
             Environment.Exit(0);
         }
 
-        // source - https://stackoverflow.com/a/2553245/12136394
         static void MoveDirectory(string source, string target)
         {
-            var files = Directory.EnumerateFiles(source, "*", SearchOption.AllDirectories)
-                                 .GroupBy(s => Path.GetDirectoryName(s));
-
+            var files = Directory.EnumerateFiles(source, "*", SearchOption.AllDirectories).GroupBy(s => Path.GetDirectoryName(s));
             foreach (var folder in files)
             {
                 var targetFolder = folder.Key.Replace(source, target);
                 Directory.CreateDirectory(targetFolder);
-
                 foreach (var file in folder)
                 {
                     var targetFile = Path.Join(targetFolder, Path.GetFileName(file));
-                    if (File.Exists(targetFile))
-                        File.Delete(targetFile);
-                    
+                    if (File.Exists(targetFile)) File.Delete(targetFile);
                     File.Move(file, targetFile);
                 }
             }
-
             Directory.Delete(source, true);
         }
     }
